Fix go tests
Expected output needed to be updated for the
int 64 as strings vs int 32 as numbers change.

Wrappers was expected for the well known types
but it is one of the ignored files. Referencing
Empty as well means we have one ignored and one
non-ignored file.

Wrapped each dir in t.Run so they will each run
even if one fails.