Add support for recursively defined message structures.
I'm trying to use this library in combination with a relatively complex
protocol, namely the remote execution protocol of the Bazel build sytem:

https://github.com/bazelbuild/remote-apis/blob/master/build/bazel/remote/execution/v2/remote_execution.proto

Some of the dependencies of this protocol contain recursively defined
messages. These currently cause the compiler to emit Elm code that isn't
valid, due to recursive alias types.

Solve this by adding a "%sMessage" custom type for every message. This
causes the recursion to be broken. Let all nested messages use this type
and fix up the encode/decode functions to properly destruct/pack
messages into the custom type.